Refactor migration steps and enhance workspace versioning
- Updated migration step file names to follow a consistent naming convention, improving clarity and organization.
- Introduced new migration steps: `migrate-00001-background-blend-filter` to normalize legacy blend mode and filter values, and `migrate-00002-seed-playgrounds` to ensure the existence of the `playgrounds` map in workspaces.
- Enhanced documentation in the migration README to reflect the new step naming and functionality, ensuring better understanding for developers.
- Updated the `CURRENT_WORKSPACE_VERSION` to align with the new migration steps, facilitating smoother transitions between workspace versions.
